Eliminate implicit conversions
Use explicit conversions everywhere. Adding bounds check assertions when
necessary.

+++ src/compton.h
@@ -38,7 +38,7 @@ enum root_flags {
 
 void add_damage(session_t *ps, const region_t *damage);
 
-long determine_evmask(session_t *ps, xcb_window_t wid, win_evmode_t mode);
+uint32_t determine_evmask(session_t *ps, xcb_window_t wid, win_evmode_t mode);
 
 xcb_window_t find_client_win(session_t *ps, xcb_window_t w);
